figma: Remove references to <crypt.h>
We need to remove references to `<crypt.h>`, which we don't have in our linux
sysroot. It looks like this was added because later versions of glibc moved
these functions into their own library. Unfortunately we run an older version
in our linux sysroot, so we don't have these headers. This was cleaned up in a
later commit to LLVM, but we just have to hack around it for now.